+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * #746 Fix default ``dateFormatter`` function to datepicker
 * #755 Fix arrow centered when table detail (thanks @wanxe)
+* #787 Fix type in horizontal field when message is empty
 
 ## 0.6.5